paths fixies after files and folders reorganizzation
