+++ src/components/creator/DetailsRecap.vue
@@ -43,7 +43,7 @@
 <script>
 import { mapGetters } from 'vuex';
 import { parseISO, format } from 'date-fns';
-import BreakpointsCond from '../../mixins/BreakpointsCond';
+import BreakpointsCond from '@/mixins/BreakpointsCond';
 
 export default {
   name: 'DetailsRecap',
