+++ src/components/creator/Stepper.vue
@@ -75,7 +75,7 @@
 </template>
 
 <script>
-import BreakpointsCond from '../../mixins/BreakpointsCond';
+import BreakpointsCond from '@/mixins/BreakpointsCond';
 
 export default {
   name: 'Stepper',
