+++ src/components/finder/filters/DateFilter.vue
@@ -113,7 +113,7 @@
 <script>
 import { format, parseISO } from 'date-fns';
 import { mapActions, mapGetters, mapMutations } from 'vuex';
-import BreakpointsCond from '../../../mixins/BreakpointsCond';
+import BreakpointsCond from '@/mixins/BreakpointsCond';
 
 export default {
   name: 'DateFilter',
