+++ src/components/finder/filters/LocationFilter.vue
@@ -95,7 +95,7 @@
 </template>
 <script>
 import { mapActions, mapGetters, mapMutations } from 'vuex';
-import BreakpointsCond from '../../../mixins/BreakpointsCond';
+import BreakpointsCond from '@/mixins/BreakpointsCond';
 
 export default {
   name: 'LocationFilter',
