+++ src/components/matchesCards/MatchCard.vue
@@ -195,7 +195,7 @@
 <script>
 import { mapMutations } from 'vuex';
 import { format, parseISO } from 'date-fns';
-import BreakpointsCond from '../../mixins/BreakpointsCond';
+import BreakpointsCond from '@/mixins/BreakpointsCond';
 
 export default {
   name: 'MatchCard',
