+++ src/components/navbar/ModeSwitcher.vue
@@ -72,7 +72,7 @@
 
 <script>
 import { mapGetters } from 'vuex';
-import BreakpointsCond from '../../mixins/BreakpointsCond';
+import BreakpointsCond from '@/mixins/BreakpointsCond';
 
 export default {
   name: 'ModeSwitcher',
