+++ src/components/navbar/Navbar.vue
@@ -96,7 +96,7 @@
 /* eslint-disable global-require */
 
 import { mapGetters, mapMutations } from 'vuex';
-import BreakpointsCond from '../../mixins/BreakpointsCond';
+import BreakpointsCond from '@/mixins/BreakpointsCond';
 import ModeSwitcher from './ModeSwitcher.vue';
 import PlayerInfo from './PlayerInfo.vue';
 
@@ -116,9 +116,9 @@ export default {
     }),
 
     getAvatarPicture() {
-      if (this.logged) {
+      if (this.user) {
         // eslint-disable-next-line import/no-dynamic-require
-        return require(`../${this.user.picture}`);
+        return require(`@/${this.user.picture}`);
       } return require('@/assets/users/match.jpg');
     },
   },
