+++ src/store/modules/auth/index.js
@@ -30,7 +30,6 @@ export default {
       const userData = res.data[0];
       if (userData) {
         commit('setUser', userData);
-        commit('setLoginStatus', true);
         return true;
       } return false;
     },
